+++ tests/testthat/test-add_phylopic_base.R
@@ -4,7 +4,7 @@ test_that("add_phylopic_base works", {
   # phylopic in background, with name
   expect_doppelganger("phylopic in background", function() {
     plot(1, 1, type = "n", main = "A cat")
-    add_phylopic_base(name = "Felis silvestris catus", ysize = .7,
+    add_phylopic_base(name = "Felis silvestris catus", height = .7,
                       verbose = TRUE)
   })
 
@@ -13,7 +13,7 @@ test_that("add_phylopic_base works", {
     cat_png <- get_phylopic("23cd6aa4-9587-4a2e-8e26-de42885004c9",
                             format = "raster")
     plot(1, 1, type = "n", main = "A cat")
-    add_phylopic_base(cat_png, x = 1, y = 1, ysize = .4, fill = "blue",
+    add_phylopic_base(cat_png, x = 1, y = 1, height = .4, fill = "blue",
                       alpha = .5, angle = -90, horizontal = TRUE)
   })
 
@@ -33,7 +33,7 @@ test_that("add_phylopic_base works", {
   expect_doppelganger("phylopics on top of plot", function() {
     plot(posx, posy, type = "n", main = "A cat herd")
     add_phylopic_base(uuid = "23cd6aa4-9587-4a2e-8e26-de42885004c9",
-                      x = posx, y = posy, ysize = sizey,
+                      x = posx, y = posy, height = sizey,
                       fill = fills, color = cols, alpha = alpha,
                       angle = angle,
                       horizontal = hor, vertical = ver)