+++ tests/testthat/test-geom_phylopic.R
@@ -10,7 +10,7 @@ test_that("geom_phylopic works", {
                             "c8f71c27-71db-4b34-ac2d-e97fea8762cf",
                             "0a8ab4f9-04c9-4485-b21a-df683d506055"))
   gg <- ggplot(df) +
-    geom_phylopic(aes(x = x, y = y, uuid = uuid), color = "purple", size = 10,
+    geom_phylopic(aes(x = x, y = y, uuid = uuid), color = "purple", height = 10,
                   horizontal = TRUE, angle = 45) +
     facet_wrap(~uuid) +
     coord_cartesian(xlim = c(1, 6), ylim = c(5, 30)) +
@@ -23,7 +23,7 @@ test_that("geom_phylopic works", {
                           format = "raster")
   gg <- ggplot(df) +
     geom_phylopic(aes(x = x, y = y), img = list(cat_png),
-                  fill = "purple", size = 10) +
+                  fill = "purple", height = 10) +
     coord_cartesian(xlim = c(1, 6), ylim = c(5, 30)) +
     theme_classic(base_size = 16)
   expect_doppelganger("geom_phylopic with png", gg)
@@ -55,7 +55,7 @@ test_that("geom_phylopic works", {
     geom_phylopic(aes(x = x, y = y), uuid = "asdfghjkl")
   expect_warning(plot(gg))
   gg <- ggplot(df) +
-    geom_phylopic(aes(x = x, y = y, uuid = uuid), size = 1E-6)
+    geom_phylopic(aes(x = x, y = y, uuid = uuid), height = 1E-6)
   expect_warning(plot(gg))
 })
 
@@ -66,7 +66,7 @@ test_that("phylopic_key_glyph works", {
   df <- data.frame(x = c(2, 4), y = c(10, 20),
                    name = c("Felis silvestris catus", "Odobenus rosmarus"))
   gg <- ggplot(df) +
-    geom_phylopic(aes(x = x, y = y, name = name, color = name), size = 10,
+    geom_phylopic(aes(x = x, y = y, name = name, color = name), height = 10,
                   show.legend = TRUE, verbose = TRUE,
                   key_glyph = phylopic_key_glyph(name = df$name)) +
     coord_cartesian(xlim = c(1, 6), ylim = c(5, 30)) +
@@ -78,7 +78,7 @@ test_that("phylopic_key_glyph works", {
 
   gg <- ggplot(df) +
     geom_phylopic(
-      aes(x = x, y = y, name = name, color = name), size = 10,
+      aes(x = x, y = y, name = name, color = name), height = 10,
       show.legend = TRUE, verbose = TRUE,
       key_glyph =
         phylopic_key_glyph(uuid = c("23cd6aa4-9587-4a2e-8e26-de42885004c9",
@@ -90,7 +90,7 @@ test_that("phylopic_key_glyph works", {
 
   cat <- get_phylopic("23cd6aa4-9587-4a2e-8e26-de42885004c9")
   gg <- ggplot(df) +
-    geom_phylopic(aes(x = x, y = y, name = name, color = name), size = 10,
+    geom_phylopic(aes(x = x, y = y, name = name, color = name), height = 10,
                   show.legend = TRUE, verbose = TRUE,
                   key_glyph = phylopic_key_glyph(img = cat)) +
     coord_cartesian(xlim = c(1, 6), ylim = c(5, 30)) +
@@ -100,7 +100,7 @@ test_that("phylopic_key_glyph works", {
   # errors/warnings
   expect_error(ggplot(df) +
                  geom_phylopic(
-                   aes(x = x, y = y, name = name, color = name), size = 10,
+                   aes(x = x, y = y, name = name, color = name), height = 10,
                    show.legend = TRUE, verbose = TRUE,
                    key_glyph =
                      phylopic_key_glyph(
@@ -112,7 +112,7 @@ test_that("phylopic_key_glyph works", {
                  theme_classic(base_size = 16))
   expect_error(ggplot(df) +
                  geom_phylopic(
-                   aes(x = x, y = y, name = name, color = name), size = 10,
+                   aes(x = x, y = y, name = name, color = name), height = 10,
                    show.legend = TRUE, verbose = TRUE,
                    key_glyph =
                      phylopic_key_glyph(name = 12345)
@@ -121,7 +121,7 @@ test_that("phylopic_key_glyph works", {
                  theme_classic(base_size = 16))
   expect_warning(ggplot(df) +
                    geom_phylopic(
-                     aes(x = x, y = y, name = name, color = name), size = 10,
+                     aes(x = x, y = y, name = name, color = name), height = 10,
                      show.legend = TRUE, verbose = TRUE,
                      key_glyph =
                        phylopic_key_glyph(name = "12345")
@@ -130,7 +130,7 @@ test_that("phylopic_key_glyph works", {
                    theme_classic(base_size = 16))
   expect_error(ggplot(df) +
                  geom_phylopic(
-                   aes(x = x, y = y, name = name, color = name), size = 10,
+                   aes(x = x, y = y, name = name, color = name), height = 10,
                    show.legend = TRUE, verbose = TRUE,
                    key_glyph =
                      phylopic_key_glyph(uuid = 12345)
@@ -139,7 +139,7 @@ test_that("phylopic_key_glyph works", {
                  theme_classic(base_size = 16))
   expect_warning(ggplot(df) +
                    geom_phylopic(
-                     aes(x = x, y = y, name = name, color = name), size = 10,
+                     aes(x = x, y = y, name = name, color = name), height = 10,
                      show.legend = TRUE, verbose = TRUE,
                      key_glyph =
                        phylopic_key_glyph(uuid = "12345")
@@ -148,7 +148,7 @@ test_that("phylopic_key_glyph works", {
                    theme_classic(base_size = 16))
   expect_error(ggplot(df) +
                  geom_phylopic(
-                   aes(x = x, y = y, name = name, color = name), size = 10,
+                   aes(x = x, y = y, name = name, color = name), height = 10,
                    show.legend = TRUE, verbose = TRUE,
                    key_glyph =
                      phylopic_key_glyph(img = 12345)